python/requirements.txt
anthropic>=0.40.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0